requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
jupyterlab
regex
dvc
tqdm 
sentence-transformers 
faiss-cpu 
langchain 
chromadb
pyarrow